+++ tests/Pim/Product/ProductServiceTest.php
@@ -10,6 +10,7 @@ namespace Wizaplace\SDK\Tests\Pim\Product;
 use GuzzleHttp\Psr7\Uri;
 use Psr\Http\Message\UriInterface;
 use Wizaplace\SDK\Exception\NotFound;
+use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
 use Wizaplace\SDK\Pagination;
 use Wizaplace\SDK\Pim\Product\CreateProductCommand;
 use Wizaplace\SDK\Pim\Product\Product;
@@ -25,7 +26,9 @@ use Wizaplace\SDK\Pim\Product\ProductListFilter;
 use Wizaplace\SDK\Pim\Product\ProductService;
 use Wizaplace\SDK\Pim\Product\ProductStatus;
 use Wizaplace\SDK\Pim\Product\ProductSummary;
+use Wizaplace\SDK\Pim\Product\Shipping;
 use Wizaplace\SDK\Pim\Product\UpdateProductCommand;
+use Wizaplace\SDK\Pim\Product\UpdateShippingCommand;
 use Wizaplace\SDK\Tests\ApiTestCase;
 
 final class ProductServiceTest extends ApiTestCase
@@ -804,6 +807,63 @@ final class ProductServiceTest extends ApiTestCase
         $this->assertNull($declinations[3]->getCrossedOutPrice());
         $this->assertNull($declinations[3]->getAffiliateLink());
         $this->assertSame('code_full_declA2', $declinations[3]->getCode());
+    }
+
+    public function testGetProductShipping()
+    {
+        $shipping = $this->buildProductService()->getShipping(5, 1);
+
+        $this->assertInstanceOf(Shipping::class, $shipping);
+    }
+
+    public function testPutProductShipping()
+    {
+        $command = new UpdateShippingCommand();
+        $command->setStatus("D")
+                ->setRates([
+                    [
+                        'amount' => 0,
+                        'value'  => 100,
+                    ],
+                    [
+                        'amount' => 1,
+                        'value'  => 50,
+                    ],
+                ])
+                ->setSpecificRate(false)
+                ->setProductId(5);
+
+        $productService = $this->buildProductService();
+
+        $productService->putShipping(1, $command);
+
+        $shipping = $productService->getShipping(5, 1);
+
+        $this->assertInstanceOf(Shipping::class, $shipping);
+        $this->assertSame(100.0, $shipping->getRates()[0]['value']);
+    }
+
+    public function testUpdateShippingCommandConstraints()
+    {
+        $command = new UpdateShippingCommand();
+        $command->setStatus("Status qui n'existe pas")
+            ->setRates([
+                [
+                    'amount' => 0,
+                    'value'  => 100,
+                ],
+                [
+                    'amount' => 1,
+                    'value'  => 50,
+                ],
+            ])
+            ->setSpecificRate(false)
+            ->setProductId(-1);
+
+        $productService = $this->buildProductService();
+
+        $this->expectException(SomeParametersAreInvalid::class);
+        $productService->putShipping(1, $command);
     }
 
     private function buildProductService($userEmail = 'admin@wizaplace.com', $userPassword = 'password'): ProductService