+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Next release
 
+- Added `getUserId` to `\Wizaplace\SDK\Organisation\OrganisationBasket`
+- Added `\Wizaplace\SDK\Catalog\Product::getSeoKeywords`
+- Fix `seoDescription` from `Wizaplace\SDK\Catalog\Product` for MVP
+- Added `\Wizaplace\SDK\Pim\Product::getShipping`
+- Added `\Wizaplace\SDK\Pim\Product::putShipping`
+- Added `\Wizaplace\SDK\Pim\UpdateShippingCommand`
+
 ### 1.37.0
 
 ### New features